requirements.txt
pandas
numpy
scikit-learn
cryptography
pyjwt
flask
flask-cors
streamlit
plotly
requests